task_2/requirements.txt
django-environ
langchain_core
langgraph
langchain_openai
uvicorn
fastapi